+++ index.html
@@ -38,6 +38,6 @@
       <div class="margin-buttom"></div>
     </section>
     
-    <script src="./app.js"></script>
+    <script type="module" src="./index.js"></script>   
   </body>
 </html>
