+++ index.js
@@ -1,5 +1,13 @@
 /* eslint max-classes-per-file: ["error", 3] */
 
+import contactPage from './modules/contact.js';
+import addBookForm from './modules/addBook.js';
+import currentTime from './modules/date.js';
+
+currentTime();
+contactPage();
+addBookForm();
+
 /* Declaration of sections */
 const titleDisplay = document.querySelector('.main-title');
 const list = document.getElementById('books-container');
@@ -9,73 +17,6 @@ const pg1 = document.querySelector('#pg1');
 const pg2 = document.querySelector('#pg2');
 const pg3 = document.querySelector('#pg3');
 let i = 0;
-
-// create add book page
-
-const formContainer = document.createElement('form');
-const bookTitleImput = document.createElement('input');
-const bookAuthorInput = document.createElement('input');
-const addBtnForm = document.createElement('button');
-// set classes
-formContainer.className = 'form-class';
-bookTitleImput.className = 'title-book';
-bookAuthorInput.className = 'author-name';
-addBtnForm.className = 'add-btn';
-// set atributes
-formContainer.setAttribute('id', 'form');
-formContainer.setAttribute('action', 'add-book');
-formContainer.setAttribute('name', 'form');
-bookTitleImput.setAttribute('id', 'title-book');
-bookTitleImput.setAttribute('type', 'text');
-bookTitleImput.setAttribute('name', 'title');
-bookTitleImput.setAttribute('placeholder', 'Title');
-bookTitleImput.setAttribute('maxlength', 60);
-bookAuthorInput.setAttribute('id', 'author-name');
-bookAuthorInput.setAttribute('type', 'text');
-bookAuthorInput.setAttribute('name', 'author');
-bookAuthorInput.setAttribute('placeholder', 'Author');
-bookAuthorInput.setAttribute('maxlength', 30);
-addBtnForm.setAttribute('type', 'button');
-addBtnForm.textContent = 'Add';
-// add book append child system
-formSection.appendChild(formContainer);
-formContainer.appendChild(bookTitleImput);
-formContainer.appendChild(bookAuthorInput);
-formContainer.appendChild(addBtnForm);
-
-// create contact page
-
-const contactParagraph = document.createElement('p');
-const contactParagraph2 = document.createElement('p');
-const contactUl = document.createElement('ul');
-const contactli1 = document.createElement('li');
-const contactli2 = document.createElement('li');
-const contactli3 = document.createElement('li');
-// contact classes
-contactParagraph.className = 'contact-text';
-contactParagraph2.className = 'contact-text';
-contactUl.className = 'contact-ul';
-contactli1.className = 'contact-li';
-contactli2.className = 'contact-li';
-contactli3.className = 'contact-li';
-// contact set atribute
-contactParagraph.setAttribute('id', 'contact-text');
-contactli1.setAttribute('id', 'contact-mail');
-contactli2.setAttribute('id', 'contact-phone');
-contactli3.setAttribute('id', 'contact-address');
-// contact append child
-contactSection.appendChild(contactParagraph);
-contactSection.appendChild(contactParagraph2);
-contactSection.appendChild(contactUl);
-contactUl.appendChild(contactli1);
-contactUl.appendChild(contactli2);
-contactUl.appendChild(contactli3);
-// contact data text
-contactParagraph.textContent = 'Do have any questions or you just want to say "Hello"?';
-contactParagraph2.textContent = 'You can reach out to us!';
-contactli1.textContent = 'Our e-mail: mail@mail.com';
-contactli2.textContent = 'Our phone number: 004-358-6534-422';
-contactli3.textContent = 'Our address: Streetname 22, 84503 City, Country.';
 
 // display main page
 
@@ -236,33 +177,3 @@ document.querySelector('#books-container').addEventListener('click', (e) => {
   // remove from store
   Store.removeBook(e.target.previousElementSibling.textContent);
 });
-
-// Clock
-
-function currentTime() {
-  const date = new Date();
-  let hh = date.getHours();
-  let mm = date.getMinutes();
-  let ss = date.getSeconds();
-  let session = 'AM';
-
-  if (hh === 0) {
-    hh = 12;
-  }
-  if (hh > 12) {
-    hh -= 12;
-    session = 'PM';
-  }
-
-  hh = hh < 10 ? `0${hh}` : hh;
-  mm = mm < 10 ? `0${mm}` : mm;
-  ss = ss < 10 ? `0${ss}` : ss;
-
-  const time = `${hh}:${mm}:${ss} ${session}`;
-
-  document.getElementById('clock').innerText = `29/06/2022, ${time}`;
-  setTimeout(() => {
-    currentTime();
-  }, 1000);
-}
-currentTime();
